fixed last awful commit and added equality and contains to Decks
fixed ValueError for only the right values.
removed a pointless line.
Decks can now be equal. this is when they have the exact same cards but not necessarily the same order.
Decks can now contain Cards ie: if card in Deck, also works for iteration
fixed a very nasty bug where shuffling randomly empties a deck if it is not the first deck to be shuffled